backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
pydantic==2.7.4
python-dateutil==2.9.0.post0
# Optional (only if you want Snowflake live connection)
snowflake-connector-python==3.11.0